blocage

+++ FRONT-END/src/app/admin/components/users/components/item-user/item-user.component.ts
@@ -1,5 +1,5 @@
-import { UserService } from './../../../../../shared/services/user/user.service';
 import { Component, Input } from '@angular/core';
+import Swal from 'sweetalert2';
 import { Users } from '../../shared/models/users';
 import { UsersService } from '../../shared/services/users.service';
 
@@ -17,6 +17,7 @@ error!: string;
 constructor(private userService: UsersService){}
 ngOnInit(){
   this.loadUsers();
+
 }
 
 loadUsers(){
@@ -28,15 +29,43 @@ loadUsers(){
     (error) => this.error = error
   )
 }
-deleteUser(id: number): void {
-  this.userService.deleteUser(id).subscribe(
-     () =>{
-       this.users = this.users.splice(id, 1)
-       this.userService.refreshNeeded.next();
-     }
-      // this.users.filter(item => item.id !== id)
-      ,
-    (error) => this.error = error
-  );
+toggleUserStatus(user: any): void {
+  if (user.statut) {
+    this.userService.deactivateUser(user._id).subscribe(response => {
+      user.statut = false;
+    });
+  } else {
+    this.userService.activateUser(user._id).subscribe(response => {
+      user.statut = true;
+    });
+  }
 }
+deleteUser(id: string): void {
+  Swal.fire({
+    title: "Etes vous sure de vouloir suppricett?",
+    text: "cet action est irreversible",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Oui, supprime le!"
+  }).then((result) => {
+    if (result.isConfirmed) {
+      this.userService.deleteUser(id).subscribe(
+        () =>{
+          // this.users = this.users.splice(id, 1)
+          // this.userService.refreshNeeded.next();
+          Swal.fire({
+            title: "Supprime !",
+            text: "Your file has been deleted.",
+            icon: "success"
+          });
+
+        });
+
+    }
+  });
+
+}
+
 }
